[refactor] Align fastapi project with full-stack-fastapi-template from tainglo repo
