+++ src/api/routes/library.py
