+++ src/api/routes/search.py
